+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Minor visual changes on badges
+
 ## 0.3.0
 
 * Introduce storage monitoring through the [Active Storage Dashboard](https://github.com/giovapanasiti/active_storage_dashboard) gem